requirements.txt
numpy
scipy
opencv-python
tqdm
lmdb
pytorch
torchvision
e3nn
matplotlib
